requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
Jinja2==3.1.2
opencv-python==4.6.0.66
Pillow==9.3.0
torch==1.13.1
transformers==4.26.0
scikit-learn==1.0.2
ffmpeg-python==0.2.0
